Add files via upload
Hello there, my project is about creating an E-Commerce webpage. I have chosen to make a webpage for Electronic gadgets. Here is a simple brief about how I have done.
DAY 1 - On day 1 I fully focused on creating an HTML page. 
DAY 2 - On day 2 tried to complete the HTML of the page fully.
DAY 3 - On day 3 created CSS to add to the HTML page.
DAY 4 - On day 4 successfully enabled the JS to give the message.